introduce overhauled ASCII serializers
On the basis of the recently introduced string serializer base, the
ASCII serialization can be overhauled by using the type visiting
system.

Because the base class does all the heavy lifting, the overhauled
implementation only has to define the corresponding traits classes.
This is trivial for the new default ASCII serializer, as it uses the
default traits class from the base implementation.

The classic single and multi line serializers from the
AsciiSerialization header on the other hand have been re-implemented by
defining corresponding traits classes.

Note that all ASCII serializers define a template parameter for a
derived class to allow additional extension via CRTP. Additionally, as
with the overhauled CBOR serializer, the overhauled ASCII serializers
do not yet replace the legacy ones and currently have to be used
explicitly.