introduce overhauled JSON serializer
Because JSON is also only another string format, JSON serialization can
be overhauled by using the string serializer base and defining a
corresponding traits class. The resulting serializer does functions the
same way as other string serializers and does not have a dependency to
any JSON library.

Note that the behavior differs slightly from the legacy version, by
serializing exactly the requested properties. If only the valid
properties are desired, they have to explicitly be specified.

Additionally, it should be mentioned that the performance of string
serialization in general is currently bottlenecked by floating point to
and from string conversion.

While the overhauled JSON serializer outperforms the legacy
implementation (which uses RapidJSON) by roughly 110% when using MSVC,
it is several times slower when using GCC. This is because MSVC fully
supports C++17 to_chars and from_chars for floating point types, while
these functions are not yet available in GCC.