remove string conversion from type descriptors
String conversion was originally included as part of the type system to
provide an easy to use way for converting values to strings in a
default format. This also included providing ostream operators for
certain types.

However, the mechanism was never complete and also resulted in string
conversion being scattered into various files and components, when it
should rather be part of serialization. Additionally, since neither the
toString functions nor the operators are parameterizable, they often do
not yield the desired results. This is particularly problematic when
the ostream operators are auto selected (e.g. in googletest).

Since the overhauled serializers provide a simple way to get the
desired format and the current descriptor format is the one of the
default string serializer, there is no reason for keeping string
conversion as part of the type system anymore.