fix incomplete host transceiver logging
The host transceiver currently does not include peer information when a
connection is closed with an error. This leads to incomplete logging
output when the logging level is set to 'notice'. Additionally, it is
missing output about new connections entirely.

Note that these changes should only be considered a workaround until
logging of the IO components is overhauled.