fix missing DotsCacheInfo descriptor for guests
Depending on how a guest application is built, it can currently happen
that the descriptor for the DotsCacheInfo type is not present in the
registry. Since that type is always sent by the host regardless of
whether there actually is a subscription for it, it always has to be
added to the local registry of a guest transceiver.