add CLI arguments for local and remote endpoints
Both the dotsd and guest application currently only support the
creation of TCP endpoints to listen on or open channels to
respectively.

With the introduction of universal endpoints, the command line
arguments can now be extended to support all available channels types
via corresponding endpoint URIs.

Note that this now also enables a user to create multiple listeners for
the dotsd to be used in the host transceiver.