restrict pack serialization to string serializers
The current way of handling pack serialization and deserialization in
the serializer base class is convoluted and error-prone. Additionally,
there do not seem to be valid use cases outside of string based
serialization anyway.

By restricting pack serialization to string serializers, it can be made
simpler it its implementation and easier to maintain.