rename pack to tuple serialization for clarity
The previous renaming from tuple to pack serialization made sense when
handling tuples was based on pack visiting from the type visitor. Since
it turned out that a different mechanism is required for handling
tuples in string serializers, the name changes to the corresponding
identifiers should be reverted to avoid confusion.