use default string settings for JSON serialization
The JSON serializer currently defaults to using compact serialization
when using the global convenience functions. Since the default
assumption when using these functions is that the user wants to obtain
a human-readable output, it should be consistent with the default
settings for string serialization instead of being a special case.