fix construction via descriptor of dynamic structs
Constructing dynamic structs via their descriptor currently always
performs the construction in-place; i.e. placing the property area
behind the address of the dynamic struct instance itself, instead of
allocating it separately. This was done under the assumption that
constructing dynamic structs via the descriptor is only done in
properties, where in-place construction is always correct (see [1]).

However, not only is this assumption error-prone, there are also
already use cases where it does not apply. Most notably, when handling
vectors of dynamic struct types, instances are currently
unintentionally constructed in-place via the descriptor. This issue
occurs with both the (removed) legacy serializers, as well as the
overhauled serializers and results in segfaults for corresponding
dynamic struct types.

To fix this issue, in-place construction has to be modeled explicitly
into the descriptor interface, so that it can be used unambiguously
only when desired (e.g. in properties).

References:

 - [1] commit bec2f5a235acbd437ec2493e1c3d9c518416d774