fix property deprecation warning issue
Even though properties can be correctly marked as deprecated
individually, the feature is currently not very viable, because some
compilers (e.g. GCC) still issue corresponding warnings, even if the
property is not being used directly. This is because the properties are
always being used indirectly, when members of the containing struct are
compiled (e.g. constructors and destructors).

This issue can be circumvented by adding a conditional warning
suppression around the struct declaration, which results in warnings
only being generated when a deprecated property is actually accessed
from outside of the header file.