add support for deprecation acknowledgment
When struct types or properties are marked as deprecated, users might
not always be able to immediately accommodate corresponding changes or
there might be other reasons why affected components cannot be changed;
e.g. to retain backwards compatibility. If a user in these situations
cares about compiler warnings they either have to manually suppress
warnings or disable deprecation warnings entirely.

Since both options are not ideal, it should be possible to acknowledge
exactly those deprecations that the user has seen and taken care of. To
do that, this commit introduces a preprocessor mechanism to disable
deprecations either for entire struct types or individual properties.