fix default compiler options for MSVC
For setting the base compiler options, the project currently checks for
the existence of the CMAKE_CXX_FLAGS variable and sets some default
values if it is not defined. The intention is to provide a mechanism by
which a user can override the default options if necessary.

However, even though this approach works when using GCC, it does not
work for MSVC, because CMake in that case always defines
CMAKE_CXX_FLAGS with some base values, which means that the intended
base values of the project are never applied. Additionally, MSVC always
issues a warning when a compiler flag gets overridden, which makes
setting of compiler options that are already part of the
CMAKE_CXX_FLAGS unfeasible.

The issue is currently being worked on and individual fixes will be
included in newer CMake versions (see [1]). Until then, the recommended
approach is to replace the values contained in the CMAKE_CXX_FLAGS as
required.

References:

 - [1] https://gitlab.kitware.com/cmake/cmake/-/issues/19084