resolve property init from incompatible literals
If an arithmetic literal is formally of an incompatible type, but does
fit into the target type, the value is usually implicitly converted
without resulting in an incompatible type warning.

However, struct property initializers currently deduce the parameter
types of their constructors. In those contexts, a call with an
arithmetic literal results in an implicit conversion of incompatible
types when for example a 'uint8' is to be initialized by literal 0
(with deduced type 'signed int'), which by some compilers (e.g. MSVC)
is remarked with a warning.

This commit resolves this issue, by providing a dedicated constructor
for arithmetic types.