simplify type introspection via descriptors
There are various use cases which require basic introspection of DOTS
types. Even though introspection is precisely the purpose of
descriptors, it is often cumbersome to do, because there is no simple
way to check the type of a descriptor or perform a necessary conversion
afterwards.

This results in a lot of manual type checks, as well as static casts to
compatible types. Using dynamic casts would include implicit type
checking, but they are generally not desirable to use, due to the
non-trivial costs.

Instead, descriptors should support basic introspection via easy to use
convenience functions. These have to support both the use case for
specific types, as well as type categories, such as struct types.