prevent implicit construction of base structs
Even though instances of the struct base class should never be created
outside of a derived struct type, the constructor has to be public,
because it is necessary to call it from outside the class for certain
dynamic scenarios.

However, the constructor should be explicit, because these kind of
instances can otherwise be created implicitly from a descriptor, which
should never be something a user would want or expect.