support registry iteration over specific types
The type registry currently only allows iteration over all contained
types. However, in many use cases, a user is only interested in certain
types or type categories (e.g. struct types).

Based on the recently introduced type introspection, the registry
iteration can be extended to accept an arbitrary number of descriptor
types, which are then used to check a descriptor before it is passed
down to the handler, casted to the explicit type. Descriptor types can
be mixed in a flexible way, as long as the handler can be called with
all provided types (e.g. by using a generic lambda).