base type subscriptions on descriptors
To acquire a type subscription, a user currently has to specify a value
for the Type enum, even though the intention is to actually subscribe
type descriptors. Additionally, this approach has a cumbersome
implementation, because each value has to be checked and mapped to the
corresponding descriptor type.

Instead, type subscriptions can be implemented in a similar way to the
recently introduced type based registry iteration (see [1]). Similar to
a struct type, a user can just specify a descriptor to subscribe, which
then gets checked and passed down to the handler, casted to the
explicit type. Similar to registry iteration, this is not limited to a
singled descriptor. Multiple descriptor types can be combined, as long
as they are accepted by the handler.

References:

 - [1] commit 3c0b06740641663b21a95d57ca443c35af083f2c