fix not null assertions in type category find
The type category find functions in the registry currently do not
incorporate the assertNotNull parameter into the type category checks.

This can lead to dereferencing of null pointers, when the type is found
in the registry, but not of the expected type; as it is the case with
the category getters.