fix possibility of conflicting types in registry
During type registration, the registry currently only considers the
non-static descriptor map to evaluate whether a type is already known.

This means, that a type with the same name as a static type can still
be registered, leading to conflicting or duplicate types during lookup
and iteration.

Instead, the registry always has to use its own lookup function to
determine whether a type is already known.