make type descriptors shared ptr only
Type descriptor instances are supposed to be unique within the type
system and generally be comparable by their address. Even though this
is currently not enforced in general, this assumption is made
particularly fragile by allowing descriptors to be constructed on the
stack and be copied.

As a first step to make their handling more robust, descriptor
construction can be made 'shared pointer only', by using a similar
approach as in the IO channels. Since descriptors are not supposed to
be created manually anyway, making construction a little more
cumbersome is a reasonable cost for making descriptors more safe to
use.