use proper const correctness for type descriptors
The constness of type descriptor usage is currently inconsistent. Even
though references are usually const, shared pointers of descriptors are
usually not. And even if they are, only const references to shared
pointers are used, which do not prevent non-const access to the managed
instances.

Since descriptors are currently not mutable anyway, this discrepancy
does not make a big difference in practice, but it might become
relevant at a later point when descriptors are allowed to be updated.
It also prevents certain refactorings that depend on const correctness
being present.

To fix this, components that manage type descriptors (e.g. the
descriptor map and type registry) have to provide proper const and
non-const access to their elements. This includes using
std::shared_ptr<const TDescriptor> when dealing with non-mutable
descriptor pointers. Using shared pointer references is generally not
preferable, because it prevents certain implicit conversions of
compatible pointer types.