simplify type descriptor shared pointer handling
Even though in most contexts it is preferable to pass type descriptors
around via references, there are also multiple contexts in which it is
required to pass a descriptor via its owning shared pointer (e.g. when
adding a new type to the registry).

Additionally, because the corresponding std::shared_ptr can currently
not be acquired from a reference, it is necessary to always have
redundant support for both references and shared pointers; resulting in
quite a lot of bloat.

However, since descriptor instance are now always guaranteed to be
created as managed by a shared pointer, descriptor handling can be
simplified by using std::enable_shared_from_this. This way, descriptors
can exclusively be passed as references, because a the managing shared
pointer can always be obtained if required.