fix rule evaluation for authentication requirement
The legacy authentication manager does currently not correctly evaluate
the list of matching authentication rules when determining whether a
network requires authentication or not.

The first issue is, that rules that do not include a secret are not
considered during evaluation. This can result in a guest being required
to authenticate themselves, even if they connect from a network that is
always accepted without a secret.

Additionally, the evaluation currently always checks all matching
rules, which defeats the purpose of having a prioritized rules list.

Finally, the default accept policy is applied in the inverted way,
resulting in an authentication requirement, even if a guest would end
up being accepted without it and vice versa.

These issues can be fixed as follows: Because guests from networks that
are not accepted currently do formally not require authentication (even
though they are rejected later), it is sufficient to find the first
matching rule with 'accept == true' and check whether a secret is
associated with it. If no such rule is found, the default accept policy
can be applied.