fix IO event issues related to 'select' syscall
On most UNIX systems, the library currently uses the 'select' system
call for IO events (e.g. for TCP sockets). This is done implicitly by
disabling 'epoll' with the BOOST_ASIO_DISABLE_EPOLL option, which
results in ASIO using 'select' as a fallback.

Even though 'select' is generally more portable than 'epoll', it
struggles with more complex applications, which use a larger number of
file descriptors, particularly in conjunction with using blocking and
non-blocking sockets simultaneously. The way ASIO is using 'select' can
result in the event loop to seemingly "stall", because calls on an open
socket suddenly return "-1 EBADF (Bad file descriptor)".

The main reason this option was used in the past, was because events on
regular files were supposedly not supported with 'epoll' in ASIO (see
[1]). However, this is not an issue with 'epoll', but all event driven
IO, including 'select' (see section 'O_NONBLOCK' in [2]).

In any case, these things are implementation details of ASIO, and
unless necessary, dots-cpp should use the implementation that ASIO
deems best for the target platform as a default. Options such as
BOOST_ASIO_DISABLE_EPOLL can still be used to override this behavior if
a user should so desire.

References:

 - [1] https://stackoverflow.com/questions/378515/whats-the-deal-with-boost-asio-and-file-i-o

 - [2] https://man7.org/linux/man-pages/man2/openat.2.html