fix false update event for subs on preloaded types
The guest transceiver usually keeps track of joined groups and only
transmits corresponding DotsMember messages if a type is subscribed for
the first time.

However, this mechanism is currently not used during preloading,
resulting in two joins of the same group being transmitted to the host,
which causes the container content to be received twice. If the
container is not empty, a user might falsely be informed of a container
update, when no instance was updated at all.