introduce base header for Google Test integration
Unit testing the library is currently done without any additional
integration of a testing framework at all. Since proper unit testing
always requires some additional components (e.g. printers and matchers)
and the tests for the library are already based on Google Test (also
known as 'googletest' or 'gtest'), it makes sense to add official
support for the framework to the library.

That way, not only the unit tests of the library itself would benefit
from such integration, but also all applications built based on the
library, which are also using Google Test themselves.

As a first step, this commit introduces a simple base header for Google
Test integration. This is not only done for convenience, but is also
necessary to ensure that all printers provided by the integration are
actually used.

For extending printers, Google Test uses either ostream operators or
PrintTo functions that are available via argument dependent lookup (see
[1]). If none are provided, a bytes printer is used as a fallback. This
has the drawback, that all translation units have to "see" the same
printers, because otherwise the first definition "wins" (which might be
the default) one (see [2]).

References:

 - [1] https://github.com/google/googletest/blob/master/docs/advanced.md#teaching-googletest-how-to-print-your-values

 - [2] https://stackoverflow.com/questions/24673515/googletest-printto-not-getting-called-for-a-class