add Google Test printer extensions for DOTS types
Since DOTS types are involved in almost all parts of the library in one
way or another, unit tests involving the library often come down to
comparing struct instances, properties or other values of DOTS types.

The library should therefore provide corresponding extensions for the
Google Test printer to make output more readable. To do this, the
printer is extended by providing an ostream template that prints DOTS
values by using the default string serialization.

Note that this has to be an ostream operator, because, for unknown
reasons, a corresponding PrintTo overload is currently not being used
as expected.