remove publisher interface
The publisher interface is currently only present for backwards
compatibility with the legacy library; primarily for unit testing.

However, the interface alone is not sufficient to provide any
meaningful ways for unit testing and is therefore currently not serving
any purpose.

Since proper Google Test integration is now part of the library, the
interface should rather be removed entirely.