add helpers for GTest call sequence expectations
One common use case for DOTS unit tests is the testing of sequences of
stimuli and reactions. Even though this is natively supported by Google
Test, it still usually takes enough additional code to be bothersome
when having to do it often.

To mitigate this, this commit introduces a generic mechanism for
constructing sequenced call expectations. This comes in two flavors: a
variant where the expectations are added to a given Google Test
sequence and another where the sequence is added consecutively within
the action of the previous expectation.

In both cases, expectations can be specified via ordered arguments,
which can be interleaved with optional actions that should be executed
when the preceding expectation is met.