add support for publish tests via Google Test
Publishing information and reacting to publish events from others is an
essential aspect of most DOTS applications. Creating corresponding
(unit) tests is currently quite complex, requires a lot of additional
code and is also tedious and error prone to do.

Instead, creating unit tests should be properly supported by the
library; with the following design goals in mind:

 - Unit testing DOTS components or applications should be easy, to
 encourage the creation of tests in the first place. The user must
 therefore be able to focus on the tests itself, rather than the
 technicalities behind them.

 - Test results must be meaningful. The user must be able to tell which
 expectation was not met and why it failed. This also includes proper
 printing of values in the test output.

 - The tests have to be representative of interactions with actual DOTS
 hosts. In particular, the user cannot be expected to mock
 implementation details of a transceiver to get representative results.

As a first step to meet above goals, this commit introduces a Google
Test base class alongside several components and macros for creating
publish tests.

The base class creates a "virtual" host transceiver and automatically
links them to guest transceivers via local channels. The class also
manages a set of mock subscribers, which can be used to create publish
expectations for individual types.

Additionally, publishes from a "virtual" guest can be spoofed to create
stimuli for test cases. Because all involved components are running in
an event loop, components can be executed in the same way they would in
the actual application, including timers, posts or other
functionalities that are dependent on an ASIO execution context.

Both creating expectations for and spoofing publish events can be done
by using a set of macros, which also retain the current source location
for proper output. Expectations can be set for either individual
publishes or publish sequences, in which case the library handles the
setting of sequencing automatically.