support posting with promise to ASIO IO context
Even though ASIO supports threadsafe posts to IO contexts from
arbitrary threads, there does not seem to be an easy way to wait for
the result with the currently used version of ASIO.

To rectify this, this commit introduces a simple convenience function,
that wraps a handler and automatically creates a promise for the
result. The function then returns a corresponding future the calling
thread can wait on.