prevent error-prone usage of property construct op
The property construct operator (i.e. the '()' operator) can currently
easily be mistaken for a getter for the property value, which can lead
to unintentional default construction of the property when it is
invalid before the call. This is particularly harmful when porting
legacy applications, where the same syntax was used for this purpose.

To mitigate this issue, this commit adds a safe-guard to the operator
to prohibit calling it without arguments. Default construction can
still be achieved by calling 'construct' explicitly.