fix order of user actions for call expectations
For consecutive call sequences, the optional user action is currently
invoked before the subsequent expectation is set. This causes the
expectation to fail, if the corresponding stimulus is created in the
user action.

Instead, the user action has to be invoked after invoking the action
that sets the subsequent expectation.