fix implicit bool conversion of property sets
Property sets can currently implicitly be converted to bool. Since bool
is an integer type, this can cause a property set to implicitly be used
in a variety of contexts, leading to unintentional and harmful results.

One example for this is the bitwise AND operation (i.e. the '&'
operator), which a user familiar with bitsets might mistake for the
intersection operator.

To prevent this, the operator must be marked 'explicit'.