allow raw string values in string deserialization
For disambiguation and escaping, string value inputs used for string
deserialization currently always have to be delimited.

However, there are several use cases (e.g. with user input), where it
is cumbersome to specify data in this form.

Instead, it should also be possible to specify a non-delimited string
when deserializing string values or properties directly (i.e. outside
of structs or vectors). In those cases, the string is used raw (i.e.
"as is"), avoiding the need for manual escaping.