fix setting of default port for TCP endpoints
The application currently automatically sets the endpoint port when it
is not set by the user.

However, the default port is only defined for TCP connections and
should not be used for other endpoint types. Additionally, for
endpoints that do not have hostnames (e.g. UNIX domain sockets), this
results in an exception, making it currently impossible to use UDS
sockets via an application argument.

Instead, the default port must only be set for TCP endpoints.