only use legacy app arguments on default endpoint
It is currently possible to mix endpoint and legacy TCP arguments in
the application, resulting in unintuitive behavior.

For example, even if a port is set within the endpoint argument, it can
get overridden by a legacy argument.

Instead, the legacy arguments should only be used when the endpoint
argument is defaulted (i.e. unspecified), so that the two methods
cannot be mixed.