generalize tuples to all serializers
String serializers currently support the serialization and
deserialization of an arbitrary number of elements with an explicit
begin and and in the form of 'tuples'.

Since this concept is applicable to all serializers, it should be
generalized so it is no longer restricted to string serializers.

To do this, tuple serialization and deserialization can be made
available via the serializer base.

Note that this adds corresponding support to the CBOR serializers,
which serialize tuples as indefinite arrays.