make type visitor visiting level user accessible
The visiting level of the type visitor is currently only used
internally to prevent recursive visiting from invoking the 'visitBegin'
and 'visitEnd' visitors.

However, knowing the current visiting level can be valuable for the
user; e.g. for indentation in a serializer.

This commit overhauls the visiting level and makes it accessible to
derived visitors. The visiting level starts at 0 and is incremented for
every nested struct and vector value that is visited.

Note that visitors that use recursive visiting are responsible for
having the correct expectations to the visiting level value.