fix deserialization of invalid escape sequences
If the input contains an escape sequence that is unknown, string
serializers currently get stuck in an endless loop, because the input
is not advanced.

Instead, invalid escape sequence must result in a corresponding token
error.

Note that this issue was not caught by the unit tests, because it was
covered up by another error.