only use value when serializing single properties
When serializing properties directly, all serializers that support
direct property serialization currently serialize them the same way as
in their struct types. This was done primarily with the intention to
allow string serializers to serialize single properties along with
their property names.

However, this use case is rare and usually a user wants to only
serialize a property value without having to check the property for
validity, which is not possible at the moment.

Additionally, the current behavior is inconsistent with
deserialization, where serializers expect the value, but not additional
info, such as the name.

For above reasons, the behavior should be changed to only output
property values. This also allows the RapidJSON serializer to support
direct property serialization, which is currently not possible.