fix subscribe and unsubscribe during dispatch
The dispatcher does currently not consider the adding or removing of
handlers during the dispatching of events or transmissions. This
results in undefined behavior when the callee changes the number of
handlers (e.g. by adding or removing subscriptions).

Not only can this lead to segmentation violations due to changes to the
container, it can also result in wrong invocations of handlers.

Depending on their id and position on the map, handlers can still be
invoked after an unsubscribe and new subscription handlers will be
invoked twice.

In general, the dispatcher has to give the following guarantees in
regards to handler invocation:

 1. A callee must always be able to safely add or remove handlers from
 within an event or transmission handler

 2. A handler must never be invoked once it has been removed

 3. A handler that is added for the same type during dispatch must not
 be called for the current event

Above points can be achieved as follows:

 - Safeguard the handler being dispatched to prevent invalidation of
 current handler iterator and remove them after dispatch

 - Use descending ordering in the handler maps to exclude new handlers
 them from being invoked for the current event

 - Everything else follows from the properties of std::map (see [1])

Note that this issue also affects descriptor subscriptions, which are
not yet fixed with this commit!

References:

 - [1] https://en.cppreference.com/w/cpp/container