properly reflect event layer in project structure
The library currently roughly consists of four layers:

 * Type layer: Provides a type system for statically and dynamically
 defined struct types

 * Types layer: Contains internally- and user-defined types based on
 the type system

 * IO layer: Provides IO abstractions via channels that transport
 (serialized) instances from the types layer in and out of an
 application

 * Event layer: Uses the IO layer to connect DOTS spaces and create
 events based on published instances of (user) types

While the first three layers are explicitly represented in their own
namespaces, the event components are currently part of the ::dots::io
namespace. Additionally, several of these components are lifted into
the ::dots namespace, when the dots.h header is included.

Not only is it confusing to have components be part of several
namespaces depending on the included hierarchy, it is also indicative
of the issue that the layers are currently not correctly reflected in
the project structure.

Since the relation of the event layer and and the other layers is
hierarchical and because events are what the user is usually interested
in, the components should be moved to the top-level in the project
hierarchy.