move registry to type namespace
Historically, the type registry had some dependencies to IO components,
was therefore part of the ::dots::io namespace and was then recently
moved to the ::dots namespace as an event component.

However, the current registry exclusively works as a registry of types
based on the type system and does not have any intrinsic connection to
IO or events. It should therefore be moved accordingly.

Note that the removal of all type:: namespace qualifiers simplifies the
Registry.cpp to such an extent that the move might not longer be
detected as a rename.