move free serialize functions to ::dots namespace
Since the purpose of the free serialize and deserialize functions is to
provide convenient access to serialization functionality, they should
be moved to the top-level ::dots namespace, so that they can be used
without the ::io qualifier.

This also removes another inconsistency when the dots.h is included,
which currently already lifts the to_string function into the ::dots
namespace.