introduce serialization layer
Mostly for historical reasons, the serialization components are
currently part of the IO layer.

However, even though serialization is usually used by the IO
components, particularly channels, serialization itself is not
intrinsically connected to the kind of IO in the library.

The corresponding components should therefore be moved into a dedicated
layer with its own namespace.