overhaul debug logging of transmissions
Direct or indirect logging of transmissions is currently scattered
around several components; often with inconsistent logging levels and
formats.

Instead, transmission logging can be performed exclusively in the
connection, which allows consistent logging in both host and guest
connections for debugging purposes.

Note that, to keep the log output clean, transmissions are now logged
for the debugging log level only.