fix UB during DotsClearCache when origin is dirty
When the origin connection of a transmit is dirty, removing it
immediately would result in undefined behavior when the receive handler
returns.

For this reason, error handling for an origin connection is currently
performed by rethrowing the exception at the end of the transmit.

However, when processing a DotsClearCache request, the necessary
removes are currently performed 'locally', which can cause a dirty
origin connection to be removed immediately, while it is still being
used. This results in undefined behavior and on Linux systems usually
to a host crash with a 'pure virtual method called' error.

Additionally, retaining the origin connection would not be sufficient,
because rethrowing the exception at the end of a transmit would
interrupt the transmitting of multiple instances, which in the case of
a DotsClearCache could result in a partially removed container.

Instead, the channel and connection receive handling can be adjusted in
such a way, so that dirty origin connection can safely be removed like
non-origin ones.