fix documentation of 'fromCache' in 'DotsHeader'
The documentation currently states that if the property is invalid or
has the value '0', the instance was not sent from cache.

However, this is false, as the property is actually exclusively set
when an instance is sent from cache and a value of '0' is used to
indicate to the guest that no more instances from cache will follow.

Note that the documentation was also already incorrect for the legacy
implementation.