fix moves of user actions in call sequences
If the user specifies an action for a call sequence, the argument
tuple can currently be moved before the action is extracted out of it.

Depending on what kind of data is part of the action (e.g. objects
captured into a lambda by value), this can result in empty or invalid
objects when the action is invoked.

To fix this, the action has to be extracted from the argument tail
tuple before the tuple is moved.