fix segfaults in publish test base destructor
The publish test base currently automatically processes all ready event
handlers in its destructor. The intention was to relieve the user of
having to explicitly take care of processing events themselves; which
is sufficient for most simple test cases.

However, this can easily result in segmentation violations, because the
derived part of the object, on which the handlers might depend on, is
already destroyed when the events are dispatched. This is particularly
treacherous, because the tests might still "work", without the user
noticing that there is an issue.

Since this is a fundamental problem, the issue can only be fixed by
removing the processing entirely.

If a user wants automatic processing, they can still easily add it to
the destructor of the derived class.