support test sequences that depend on self updates
The mock handler of a publish expectation is currently invoked when the
transmission is dispatched by the host. This makes it impossible to
create a stimulus from a publish expectation action that depends on a
corresponding self update of the guest container. This is true, even if
the creation is delayed with a post to the IO context.

This can be resolved, by delaying the invocation of the mock handler,
until the self update has already been queued for execution.

Note that the mock handler will still intentionally be invoked before
the self update. Otherwise it would prevent the creation of stimuli or
expectations that depend on the self update event itself.