fix memory leak for dynamic AnyStruct instances
Creating an AnyStruct with a dynamic struct type currently always
results in leaked memory, because the property area of the dynamic
struct is never released.

This is a regression that was introduced when the explicit in-place
construction was added (see [1]), which unintentionally and implicitly
changed the AnyStruct to use "out-of-place" allocation, while using
in-place construction before.

References:

 - [1] commit bc770aa8af821ad63ed6178a23d0abd3cafa9ac9