prevent unintentional calls on temporary container
Due to the similarity of the 'Container' class name and the global
'container' accessor, a user might create a temporary container object
instead of accessing a global container.

Since this is a completely valid operation in C++, corresponding issues
will usually not be detected at compile time.

To prevent this, all operations on temporary container objects can be
explicitly deleted.

Note that this does not impact the functionality, since a (default)
constructed container is always empty.